Sunny/9999/wysiwyg/requirements.txt
alembic==1.4.2
atomicwrites==1.4.0
attrs==19.3.0
boto3==1.13.21
botocore==1.16.21
click==7.1.2
colorama==0.4.3
docutils==0.15.2
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pycryptodome==3.9.7
pyparsing==2.4.7
pytest==5.4.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
s3transfer==0.3.3
secret==0.8
six==1.15.0
SQLAlchemy==1.3.17
tabulate==0.8.7
urllib3==1.25.9
wcwidth==0.2.3
Werkzeug==1.0.1
WTForms==2.3.1
